@@ -3,7 +3,7 @@ name = "gcal-to-discord"
 version = "0.1.0"
 description = "Sync Google Calendar events to Discord channel messages"
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.14"
 dependencies = [
     "google-auth>=2.35.0",
     "google-auth-oauthlib>=1.2.0",
@@ -33,7 +33,7 @@ build-backend = "hatchling.build"
 
 [tool.ruff]
 line-length = 100
-target-version = "py311"
+target-version = "py314"
 
 [tool.ruff.lint]
 select = [
@@ -50,7 +50,7 @@ ignore = [
 ]
 
 [tool.mypy]
-python_version = "3.11"
+python_version = "3.14"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
